%%File daily2weekly.Rd
%% Part of the hydroTSM R package, http://www.rforge.net/hydroTSM/ ; 
%%                                 https://CRAN.R-project.org/package=hydroTSM
%% Copyright 2023-2023 Mauricio Zambrano-Bigiarini
%% Distributed under GPL 2 or later

\name{daily2weekly}
\Rdversion{1.1}
\alias{daily2weekly}
\alias{daily2weekly.default}
\alias{daily2weekly.zoo}
\alias{daily2weekly.data.frame}
\alias{daily2weekly.matrix}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Daily -> Weekly
}
\description{
Generic function for transforming a DAILY (or sub-daily) regular time series into a WEEKLY one
}
\usage{
daily2weekly(x, ...)

\method{daily2weekly}{default}(x, FUN, na.rm=TRUE, na.rm.max=0,
        week.date.format = "\%Y-\%W", week.grouping = "calendar", ...)

\method{daily2weekly}{zoo}(x, FUN, na.rm=TRUE, na.rm.max=0,
        week.date.format = "\%Y-\%W", week.grouping = "calendar", ...)

\method{daily2weekly}{data.frame}(x, FUN, na.rm=TRUE, na.rm.max=0, dates=1, 
        date.fmt = "\%Y-\%m-\%d", week.date.format = "\%Y-\%W",
        week.grouping = "calendar", out.type = "data.frame",
        out.fmt="numeric", verbose=TRUE, ...)

\method{daily2weekly}{matrix}(x, FUN, na.rm=TRUE, na.rm.max=0, dates=1, 
        date.fmt = "\%Y-\%m-\%d", week.date.format = "\%Y-\%W",
        week.grouping = "calendar", out.type = "data.frame",
        out.fmt="numeric", verbose=TRUE, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
zoo, data.frame or matrix object, with (sub)daily time series. \cr
Measurements at several gauging stations can be stored in a data.frame or matrix object, and in that case, each column of \code{x} represents the time series measured in each gauging station, and the column names of \code{x} have to correspond to the ID of each station (starting by a letter).
}
  \item{FUN}{
Function that have to be applied for transforming from daily to weekly time step (e.g., for precipitation \code{FUN=sum} and for temperature and streamflow ts \code{FUN=mean}). \cr

\code{FUN} MUST accept the \code{na.rm} argument, because \code{na.rm} is passed to \code{FUN}.
}
  \item{na.rm}{
Logical. Should missing values be removed? \cr
-) \kbd{TRUE} : the weekly values are computed only for weeks with a percentage of missing values less than \code{na.rm.max} \cr
-) \kbd{FALSE}: if there is AT LEAST one NA within a month, the corresponing weekly values in the output object will be \code{NA}.
}
  \item{na.rm.max}{
Numeric in [0, 1]. It is used to define the maximum percentage of missing values allowed in each week to keep the weekly aggregated value in the output object of this function. In other words, if the percentage of missing values in a given week is larger than \code{na.rm.max} the corresponding weekly value will be \code{NA}.
}
  \item{dates}{
numeric, factor or Date object indicating how to obtain the dates for each gauging station \cr
If \code{dates} is a number (default), it indicates the index of the column in \var{x} that stores the dates \cr
If \code{dates} is a factor, it is converted into Date class, using the date format specified by \code{date.fmt}  \cr
If \code{dates} is already of Date class, the code verifies that the number of days on it be equal to the number of elements in \code{x}
}
  \item{date.fmt}{
character indicating the format in which the dates are stored in \var{dates}, e.g. \kbd{\%Y-\%m-\%d}. See \code{format} in \code{\link[base]{as.Date}}. \cr
ONLY required when \code{class(dates)=="factor"} or \code{class(dates)=="numeric"}.
}
  \item{week.date.format}{
character indicating how weekly dates are stored in the output object. Valid values are: \cr
-) \kbd{\%Y-\%W}: year and Monday-based week number, preserving the historical behaviour. \cr
-) \kbd{\%Y-\%m-\%d}: date of the first day available in each weekly aggregation group, returned as a \code{Date} index in zoo outputs and displayed as \kbd{YYYY-MM-DD}.
}
  \item{week.grouping}{
character indicating how daily values are grouped into weekly values. Valid values are: \cr
-) \kbd{calendar}: calendar year-week groups defined by \kbd{\%Y-\%W}, preserving the historical behaviour. Weeks crossing a calendar-year boundary are split. \cr
-) \kbd{sequential}: consecutive 7-day groups starting at the first date in \code{x}, equivalent to \code{seq(start(x), end(x), by="weeks")} for a complete daily series.
}
  \item{out.type}{
Character that defines the desired type of output. Valid values are:  \cr
-) \kbd{data.frame}: a data.frame, with as many columns as stations are included in \code{x}, and row names indicating the weekly label for each value. \cr
-) \kbd{db}        : a data.frame, with 4 columns will be produced. \cr
                 The first column (StationID) stores the ID of the station, \cr
                 The second column (Year) stores the year \cr
                 The third column (Week) stores the week number when \code{week.date.format="\%Y-\%W"} and the weekly date when \code{week.date.format="\%Y-\%m-\%d"}  \cr
                 The fourth column (Value) stores the numerical values corresponding to the values specified in the three previous columns. \cr
}
  \item{out.fmt}{
OPTIONAL. Only used when \code{x} is a matrix or data.frame object /cr
character, for selecting if the result will be a matrix/data.frame or a zoo object. Valid values are: \kbd{numeric}, \kbd{zoo}.
}
  \item{verbose}{
logical; if TRUE, progress messages are printed 
}
  \item{\dots}{
arguments additional to \code{na.rm} passed to \code{FUN}.
}
}
\details{
The argument \code{week.grouping} defines how observations are assigned to
weekly aggregation groups before \code{FUN} is applied. It is independent from
\code{week.date.format}, which only controls how the weekly groups are labelled
in the output object.

When \code{week.grouping="calendar"} (default), \code{daily2weekly} groups the
input dates using \code{format(time(x), "\%Y-\%W")}. This is the historical
behaviour of the function. The \kbd{\%W} conversion uses Monday as the first day
of the week and numbers weeks within each calendar year. Days before the first
Monday of a calendar year belong to week \code{"00"}. Therefore, a Monday-Sunday
week that crosses from December into January can be split into two aggregation
groups, one for the ending calendar year and one for week \code{"00"} or
\code{"01"} of the new calendar year.

When \code{week.grouping="sequential"}, \code{daily2weekly} groups the input
dates into consecutive 7-day periods starting at the first date in \code{x}.
For a complete daily series, the resulting weekly labels with
\code{week.date.format="\%Y-\%m-\%d"} correspond to
\code{seq(start(x), end(x), by="weeks")}. This option is useful when the user
wants uninterrupted 7-day blocks over the full record instead of calendar
year-week groups.

The argument \code{week.date.format} controls only the time index returned by
\code{daily2weekly}. With \code{week.date.format="\%Y-\%W"}, the output is
labelled with year-week strings. With
\code{week.date.format="\%Y-\%m-\%d"}, the output is labelled with the first
date of each weekly group. The missing-value filter controlled by
\code{na.rm.max} is applied to the same groups defined by \code{week.grouping}.
}
\value{
a zoo object with weekly time frequency
}
%%\references{
%% ~put references to the literature/web site here ~
%%}
\author{
Mauricio Zambrano-Bigiarini, \email{mzb.devel@gmail}
}
%%\note{
%%  ~~further notes~~
%%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{cmv}}, \code{\link{subhourly2hourly}}, \code{\link{daily2monthly}}, \code{\link{daily2annual}}, \code{\link{subdaily2daily}}, \code{\link{weeklyfunction}}, \code{\link{hydroplot}}, \code{\link{vector2zoo}}, \code{\link{izoo2rzoo}}, \code{\link{as.Date}}
}
\examples{
######################
## Ex1: Computation of weekly values, removing any missing value in 'x'

# Loading the DAILY precipitation data at SanMartino
data(SanMartinoPPts)
x <- SanMartinoPPts

# Subsetting 'x' to its first three weeks (Jan/1921 - Mar/1921)
x <- window(x, end="1921-03-31")

## Transforming into NA the 10% of values in 'x'
set.seed(10) # for reproducible results
n           <- length(x)
n.nas       <- round(0.1*n, 0)
na.index    <- sample(1:n, n.nas)
x[na.index] <- NA

## Agreggating from Daily to Weekly, removing any missing value in 'x'
w <- daily2weekly(x, FUN=sum, na.rm=TRUE)

## Same aggregation, with the weekly zoo index shown as YYYY-MM-DD dates
w.dates <- daily2weekly(x, FUN=sum, na.rm=TRUE, week.date.format="\%Y-\%m-\%d")

## Aggregation using consecutive 7-day groups starting at the first date in 'x'
w.seq <- daily2weekly(x, FUN=sum, na.rm=TRUE, week.grouping="sequential",
                      week.date.format="\%Y-\%m-\%d")

######################
## Ex2: Computation of Weekly values only when the percentage of NAs in each
#       week is lower than a user-defined percentage (10% in this example).

# Loading the DAILY precipitation data at SanMartino
data(SanMartinoPPts)
x <- SanMartinoPPts

# Subsetting 'x' to its first three weeks (Jan/1921 - Mar/1921)
x <- window(x, end="1921-03-31")

## Transforming into NA the 10% of values in 'x'
set.seed(10) # for reproducible results
n           <- length(x)
n.nas       <- round(0.1*n, 0)
na.index    <- sample(1:n, n.nas)
x[na.index] <- NA

## Daily to Weekly, only for weeks with less than 10% of missing values
w2 <- daily2weekly(x, FUN=sum, na.rm=TRUE, na.rm.max=0.1)

# Verifying that the weeks 01, 02, 06, 08, 10, 11, 12 of 'x' had 10% or more of missing values
cmv(x, tscale="weekly")

######################
## Ex3: Computation of Weekly values in a two-column zoo object, 
##      only when the percentage of NAs in each week is lower than a user-defined 
##      percentage (10% in this example).

# Loading the DAILY precipitation data at SanMartino
data(SanMartinoPPts)
x <- SanMartinoPPts

# Subsetting 'x' to its first three weeks (Jan/1921 - Mar/1921)
x <- window(x, end="1921-03-31")

## Transforming into NA the 10% of values in 'x'
set.seed(10) # for reproducible results
n           <- length(x)
n.nas       <- round(0.1*n, 0)
na.index    <- sample(1:n, n.nas)
x[na.index] <- NA

## Creating a two-column zoo object
X <- cbind(x, y=x)

## Daily to Weekly, only for weeks with less than 10% of missing values
w2 <- daily2weekly(X, FUN=sum, na.rm=TRUE, na.rm.max=0.1)

# Verifying that the weeks 01, 02, 06, 08, 10, 11, 12 of 'x' had 10% or more of missing values
cmv(X, tscale="weekly")
}



% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{manip}
%%\keyword{ ~kwd2 }
