\name{EbroDEM1000m}
\docType{data}
\alias{EbroDEM1000m}
\title{
Digital Elevation Model (DEM) of the Ebro River basin (Spain)
}
\description{
Digital Elevation Model of the Ebro River basin (Spain), with cells of 1000x1000 [m] 
}
\usage{
data(EbroDEM1000m)
}

\details{
Digital elevation model of the Ebro River basin, with 1000x1000m of spatial resolution.
}
\format{
'SpatialGridDataFrame'
}
\source{
The original file (Modelo Digital MDT200 RASTER) was downloaded from the web site of the Confederacion Hidrografica del Ebro (CHE) \url{http://www.chebro.es/} (original link http://oph.chebro.es/ContenidoCartoApoyo.htm, last accessed [March 2010]), and then resampled up to 1000x1000m of resolution. \cr

This dataset is intended to be used for research purposes only, being distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY.
}

%%\references{
%% ~put references to the literature/web site here ~
%%}
\keyword{datasets}
