\name{EbroPPtsMonthly}
\docType{data}
\alias{EbroPPtsMonthly}
\title{
Ebro Monthly Precipitation Time Series
}
\description{
Time series of monthly precipitation on 331 stations of the Ebro River basin (NW Spain), for the period January/1961 to December/1963.
}
\usage{
data(EbroPPtsMonthly)
}

\details{
Monthly time series of precipitation on 331 stations of the Ebro River basin (Spain), where some data were in-filled by using the MOSS method and from daily time series used in the technical report "Estudio de Recursos de la Cuenca del Ebro".
}
\format{
A data.frame with 331 monthly time series, plus the first field storing the dates corresponding to each row.
}
\source{
Downloaded from the web site of the Confederacion Hidrografica del Ebro (CHE) \url{http://www.chebro.es/} (original link http://oph.chebro.es/DOCUMENTACION/PrecipitacionMensualRelleno/PrecipitacionMensualRelleno.html, last accessed [March 2010]). \cr
These data are intended to be used for research purposes only, being distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY.
}

%%\references{
%% ~put references to the literature/web site here ~
%%}
\keyword{datasets}
