\name{hydroTSM-internal}
\title{Internal hydroTSM objects}
\alias{.che2zoo}
\alias{.CHE_daily_analysis}
\alias{.DailyTS2DB}
\alias{.dwi.che}
\alias{.hydroplotboxplot}
\alias{.hydroplothist}
\alias{.hydroplotts}
\alias{.northarrow}
\alias{.plot.daily.analysis}
\alias{.Qcsv2zoo}
\alias{.scalebar}
\alias{.dmc}
\alias{.drty.dcs2zoo}
\alias{.interpol1}
\alias{check_from_and_to_and_subset}
\description{Internal hydroTSM objects.}
\details{These are not to be called by the user.}
\keyword{internal}
