\name{OcaEnOnaQts}
\docType{data}
\alias{OcaEnOnaQts}
\title{
Oca in "Ona" (Q0931), time series of daily streamflows.
}
\description{
Time series with daily streamflows of the Oca River (subcatchment of the Ebro River basin, Spain) measured at the gauging station "Ona" (Q093), for the period 01/Jan/1961 to 31/Dic/1963
}
\usage{
data(OcaEnOnaQts)
}

%%\details{
%%
%%}
\format{
zoo object.
}
\source{
Downloaded from the web site of the Confederacion Hidrografica del Ebro (CHE) \url{http://oph.chebro.es/documentacion/CaudalEA/CaudalEA.htm}. Last accessed [March 2010]. \cr

These data are intended to be used for research purposes only, being distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY.
}

%%\references{
%% ~put references to the literature/web site here ~
%%}
\keyword{datasets}
