\name{KarameaAtGorgeQts}
\docType{data}
\alias{KarameaAtGorgeQts}
\title{
Karamea at Gorge, time series of hourly streamflows
}
\description{
Time series with hourly streamflows for the Karamea River(New Zeland) measured at the gauging station "Gorge", for the period 01/Jan/1980 to 31/Dec/1985.
Station Number: 95102, Easting Coordinate (NZMG): 2444629.0, Northing Coordinate (NZMG): 5994427.0, Catchment Area (km2): 1160.0.
}
\usage{
data(KarameaAtGorgeQts)
}

%%\details{
%%
%%}
\format{
zoo object.
}
\source{
Provided by the National Institute of Water and Atmospheric Research \url{http://www.niwa.co.nz/}, thanks to the gentle collaboration of Shailesh Singh \cr

These data are intended to be used for research purposes only, being distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY.
}

%%\references{
%% ~put references to the literature/web site here ~
%%}
\keyword{datasets}
